+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/Common.kt
@@ -12,7 +12,6 @@ import kotlin.jvm.JvmInline
 typealias Identifier = Long
 typealias MessageIdentifier = MessageId
 typealias UpdateIdentifier = Long
-typealias MediaGroupIdentifier = String
 typealias ForwardSignature = String
 typealias ForwardSenderName = String
 typealias AuthorSignature = ForwardSignature