+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * `MessageId` now is `value class`
   * `MessageThreadId` now is `value class`
   * `InlineQueryIdentifier` now is `value class`
+  * `MediaGroupIdentifier` now is `value class`
 
 ## 10.1.2